[webpack]: WASM support and DefinePlugin NODE_ENV
* enable loading WASM using webpack 5's inbuild wasm support

* construct URL based on that

* allows runtime checking of build (dev/prod) for loading service worker
  and webassembly
