--- src/loadServiceWorker.ts
@@ -1,16 +1,0 @@
-const BASE_URL = window?.location?.pathname ?? '/';
-
-export const loadServiceWorker = () => {
-    if ('serviceWorker' in navigator) {
-        window.addEventListener('load', () => {
-            navigator.serviceWorker
-                .register(`${BASE_URL}service-worker.js`)
-                .then((registration) => {
-                    console.log('SW registered: ', registration);
-                })
-                .catch((registrationError) => {
-                    console.log('SW registration failed: ', registrationError);
-                });
-        });
-    }
-};
