+++ src/react-app-env.d.ts
@@ -5,3 +5,5 @@ declare module '*.svg' {
     const content: string;
     export default content;
 }
+
+declare module '*.wasm';
